[BUGFIX] Fix phpstan errors from TypesAssignedToPropertiesRule
* EXT:dashboard
* EXT:extbase
* EXT:extensionmanager
* EXT:filelist

Resolves: #94124
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69320
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>

+++ typo3/sysext/extensionmanager/Classes/Domain/Repository/BulkExtensionRepositoryWriter.php
@@ -146,7 +146,7 @@ class BulkExtensionRepositoryWriter implements \SplObserver
         $countOfBindParamsPerRow = count(self::$fieldNames);
         // flush at least chunks of 50 elements - in case the currently used
         // database platform does not support that, the threshold is lowered
-        $this->maxRowsPerChunk = min(
+        $this->maxRowsPerChunk = (int)min(
             $this->maxRowsPerChunk,
             floor($maxBindParameters / $countOfBindParamsPerRow)
         );
